+++ src/main/java/com/skycatdev/antiscan/mixin/ServerHandshakeNetworkHandlerMixin.java
@@ -20,13 +20,13 @@ public abstract class ServerHandshakeNetworkHandlerMixin {
 
     //? if >=1.20.5 {
     @WrapMethod(method = "login")
-    private void antiScan$tarpitBaddies(HandshakeC2SPacket packet, boolean transfer, Operation<Void> original) {
+    private void antiScan$handleBaddies(HandshakeC2SPacket packet, boolean transfer, Operation<Void> original) {
         Utils.handleIpConnection(AntiScan.CONFIG.getHandshakeMode(), AntiScan.CONFIG.getHandshakeAction(), AntiScan.CONFIG.isHandshakeReport(), connection, () -> original.call(packet, transfer));
     }
     //?}
     //? if <1.20.5 {
     /*@WrapMethod(method = "onHandshake")
-    private void antiScan$tarpitBaddies(HandshakeC2SPacket packet, Operation<Void> original) {
+    private void antiScan$handleBaddies(HandshakeC2SPacket packet, Operation<Void> original) {
         Utils.handleIpConnection(AntiScan.CONFIG.getHandshakeMode(), AntiScan.CONFIG.getHandshakeAction(), AntiScan.CONFIG.isHandshakeReport(), connection, () -> original.call(packet));
     }
     *///?}