+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 - Fix failure to launch on the second try
+- Renaming tarpit to timeout